create tp with spliting tp1
