+++ Pipfile
@@ -9,4 +9,4 @@ pytest = "*"
 [dev-packages]
 
 [requires]
-python_version = "3.9"
+python_version = "3.10"
